Move utils to subpackage
